+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 ## [Unreleased]
 
 ### Features
+- **assert:** added AssertDirEmpty and AssertDirNotEmpty. fixes [#37](https://github.com/MarvinJWendt/testza/issues/37) ([#52](https://github.com/MarvinJWendt/testza/issues/52))
+- **assert:** added `AssertFileExists` and `AssertNoFileExists`
+
+
+<a name="v0.2.9"></a>
+## [v0.2.9] - 2021-09-21
+### Code Refactoring
+- **assert:** renamed internal variable
+- **assert:** moved `AssertCompareHelper` to assertion_helper.go
+- **assert:** moved getter functions to internal package
+- **internal:** moved `AssertRegexpHelper` to right file
+
+
+<a name="v0.2.8"></a>
+## [v0.2.8] - 2021-09-19
+### Features
 - **assert:** added `AssertRegexp` and `AssertNotRegexp`
 - **assert:** added `AssertIncreasing` and `AssertDecreasing`
 - **assert:** added `AssertLen`
@@ -190,7 +206,9 @@ Functions have a new structure.
 - **internal:** rewrite `Fail` for a nicer output
 
 
-[Unreleased]: https://github.com/MarvinJWendt/testza/compare/v0.2.7...HEAD
+[Unreleased]: https://github.com/MarvinJWendt/testza/compare/v0.2.9...HEAD
+[v0.2.9]: https://github.com/MarvinJWendt/testza/compare/v0.2.8...v0.2.9
+[v0.2.8]: https://github.com/MarvinJWendt/testza/compare/v0.2.7...v0.2.8
 [v0.2.7]: https://github.com/MarvinJWendt/testza/compare/v0.2.6...v0.2.7
 [v0.2.6]: https://github.com/MarvinJWendt/testza/compare/v0.2.5...v0.2.6
 [v0.2.5]: https://github.com/MarvinJWendt/testza/compare/v0.2.4...v0.2.5